feat: adicionando funcionalidade de remoção de endereço, caso o usuário seja removido do sistema

+++ hibisco-ws/src/main/java/enterprise/hibisco/hibiscows/service/HospitalService.java
@@ -120,7 +120,11 @@ public class HospitalService {
     }
 
     public ResponseEntity<?> deleteHospital(Long idUser) {
-        if (repository.existsById(idUser)) {
+        Long idAddress;
+        Optional<Long> findIdAddress = repository.findFkAddressByIdHospital(idUser);
+        if (repository.existsById(idUser) && findIdAddress.isPresent()) {
+            idAddress = findIdAddress.get();
+            addressRepository.deleteById(idAddress);
             repository.deleteById(idUser);
             return ResponseEntity.status(HttpStatus.OK).build();
         }
